Add fast approximation of "most interior" point
We now find the Chebyshev center of (up to) 50 randomly sampled
halfspaces' intersection rather than all of them at once. This should be
sufficient, especially since we expect almost all of the halfspaces to
be redundant, similar, and/or represent non-admissible partitions.

If this computation fails or returns a point not clearly in the interior
of all halfspaces, we fall back to the simpler approach used earlier.